createAccessToken complete fct
